requirements.txt
google-cloud-pubsub==2.21.5
# fastapi==0.111.0